Rename TS110F_dimmer to ZB006-X. https://github.com/Koenkk/zigbee-herdsman-converters/commit/c5d6f372c7b9ee6f455e62f858742a25d91a6171
